v0.10.1

* tag 'v0.10.1':
  Translation update
  Fix monitor deadlock on panic
  Folders queried by the REST interface might not exist in the model (fixes #823)
  only convert maxAge from days to seconds once
  Introducer should default to false (fixes #825)
  Cleanup the code in internal/files/leveldb.go a bit.
  Emit info level log when deleting directory fails
  Refactor out functions in main
  Reinstate ignore permissions
  ldbGenericReplace: correctly handle files with same version but different flags.
  Print some debug info when panicing in leveldb
  Slightly improve panic log format (...)
  Restarting monitor process is still considered a restart (fixes #809)
  Better version change detection (fixes #808, fixes #810)

+++ internal/model/puller.go
@@ -62,12 +62,13 @@ var (
 )
 
 type Puller struct {
-	folder    string
-	dir       string
-	scanIntv  time.Duration
-	model     *Model
-	stop      chan struct{}
-	versioner versioner.Versioner
+	folder      string
+	dir         string
+	scanIntv    time.Duration
+	model       *Model
+	stop        chan struct{}
+	versioner   versioner.Versioner
+	ignorePerms bool
 }
 
 // Serve will run scans and pulls. It will return when Stop()ed or on a
@@ -318,6 +319,9 @@ func (p *Puller) pullerIteration(ncopiers, npullers, nfinishers int) int {
 func (p *Puller) handleDir(file protocol.FileInfo) {
 	realName := filepath.Join(p.dir, file.Name)
 	mode := os.FileMode(file.Flags & 0777)
+	if p.ignorePerms {
+		mode = 0755
+	}
 
 	if debug {
 		curFile := p.model.CurrentFolderFile(p.folder, file.Name)
@@ -339,17 +343,17 @@ func (p *Puller) handleDir(file protocol.FileInfo) {
 			if err = osutil.InWritableDir(mkdir, realName); err == nil {
 				p.model.updateLocal(p.folder, file)
 			} else {
-				l.Infof("Puller (folder %q, file %q): %v", p.folder, file.Name, err)
+				l.Infof("Puller (folder %q, dir %q): %v", p.folder, file.Name, err)
 			}
 			return
 		}
 
 		// Weird error when stat()'ing the dir. Probably won't work to do
 		// anything else with it if we can't even stat() it.
-		l.Infof("Puller (folder %q, file %q): %v", p.folder, file.Name, err)
+		l.Infof("Puller (folder %q, dir %q): %v", p.folder, file.Name, err)
 		return
 	} else if !info.IsDir() {
-		l.Infof("Puller (folder %q, file %q): should be dir, but is not", p.folder, file.Name)
+		l.Infof("Puller (folder %q, dir %q): should be dir, but is not", p.folder, file.Name)
 		return
 	}
 
@@ -357,10 +361,12 @@ func (p *Puller) handleDir(file protocol.FileInfo) {
 	// don't handle modification times on directories, because that sucks...)
 	// It's OK to change mode bits on stuff within non-writable directories.
 
-	if err := os.Chmod(realName, mode); err == nil {
+	if p.ignorePerms {
+		p.model.updateLocal(p.folder, file)
+	} else if err := os.Chmod(realName, mode); err == nil {
 		p.model.updateLocal(p.folder, file)
 	} else {
-		l.Infof("Puller (folder %q, file %q): %v", p.folder, file.Name, err)
+		l.Infof("Puller (folder %q, dir %q): %v", p.folder, file.Name, err)
 	}
 }
 
@@ -370,6 +376,8 @@ func (p *Puller) deleteDir(file protocol.FileInfo) {
 	err := osutil.InWritableDir(os.Remove, realName)
 	if err == nil || os.IsNotExist(err) {
 		p.model.updateLocal(p.folder, file)
+	} else {
+		l.Infof("Puller (folder %q, dir %q): delete: %v", p.folder, file.Name, err)
 	}
 }
 
@@ -509,14 +517,16 @@ func (p *Puller) handleFile(file protocol.FileInfo, copyChan chan<- copyBlocksSt
 // thing that has changed.
 func (p *Puller) shortcutFile(file protocol.FileInfo) {
 	realName := filepath.Join(p.dir, file.Name)
-	err := os.Chmod(realName, os.FileMode(file.Flags&0777))
-	if err != nil {
-		l.Infof("Puller (folder %q, file %q): shortcut: %v", p.folder, file.Name, err)
-		return
+	if !p.ignorePerms {
+		err := os.Chmod(realName, os.FileMode(file.Flags&0777))
+		if err != nil {
+			l.Infof("Puller (folder %q, file %q): shortcut: %v", p.folder, file.Name, err)
+			return
+		}
 	}
 
 	t := time.Unix(file.Modified, 0)
-	err = os.Chtimes(realName, t, t)
+	err := os.Chtimes(realName, t, t)
 	if err != nil {
 		l.Infof("Puller (folder %q, file %q): shortcut: %v", p.folder, file.Name, err)
 		return
@@ -643,11 +653,13 @@ func (p *Puller) finisherRoutine(in <-chan *sharedPullerState) {
 			}
 
 			// Set the correct permission bits on the new file
-			err = os.Chmod(state.tempName, os.FileMode(state.file.Flags&0777))
-			if err != nil {
-				os.Remove(state.tempName)
-				l.Warnln("puller: final:", err)
-				continue
+			if !p.ignorePerms {
+				err = os.Chmod(state.tempName, os.FileMode(state.file.Flags&0777))
+				if err != nil {
+					os.Remove(state.tempName)
+					l.Warnln("puller: final:", err)
+					continue
+				}
 			}
 
 			// Set the correct timestamp on the new file
